req1/requirements.txt
pandas
psycopg2
PyYAML